chore: restructure and export
