+++ src/index.js
@@ -324,6 +324,10 @@ export {
 
 export { renderValue as formatValue } from './modules/renderValue.js'
 
+// Modules: response
+
+export { transformResponse as transformEventAggregateResponse } from './modules/response/event/response.js'
+
 // Utils: colorSets
 export {
     COLOR_SET_DEFAULT,
