+++ src/modules/response/event/__tests__/default.spec.js
@@ -1,14 +1,14 @@
-import dateResponseHideNa from '../../../__demo__/data/event/date.data.hidena.json'
-import dateResponseOrg from '../../../__demo__/data/event/date.data.org.json'
-import datetimeResponseHideNa from '../../../__demo__/data/event/datetime.data.hidena.json'
-import datetimeResponseOrg from '../../../__demo__/data/event/datetime.data.org.json'
-import numericResponseHideNa from '../../../__demo__/data/event/integer.data.hidena.json'
-import numericResponseOrg from '../../../__demo__/data/event/integer.data.org.json'
+import dateResponseHideNa from '../../../../__demo__/data/event/date.data.hidena.json'
+import dateResponseOrg from '../../../../__demo__/data/event/date.data.org.json'
+import datetimeResponseHideNa from '../../../../__demo__/data/event/datetime.data.hidena.json'
+import datetimeResponseOrg from '../../../../__demo__/data/event/datetime.data.org.json'
+import numericResponseHideNa from '../../../../__demo__/data/event/integer.data.hidena.json'
+import numericResponseOrg from '../../../../__demo__/data/event/integer.data.org.json'
 import {
     VALUE_TYPE_DATE,
     VALUE_TYPE_DATETIME,
     VALUE_TYPE_NUMBER,
-} from '../../valueTypes.js'
+} from '../../../valueTypes.js'
 import {
     applyDefaultHandler,
     getDimensions,
